requirements.txt
pandas==2.1.0
apache-airflow==2.7.1
python-decouple==3.8
azure-storage-blob==12.18.2